+++ test/jest/__mock__/index.js
@@ -1,0 +1,5 @@
+import './currencyData.mock';
+import './matchMedia.mock';
+import './stripesConfig.mock';
+import './stripesCore.mock';
+import './stripesIcon.mock';