+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change history for ui-plugin-find-import-profile
 
-## **5.0.0** (in progress)
+## **5.1.0** (in progress)
+
+### Features added:
+* Cover `<FindImportProfile>` component with tests (UIPFIMP-42)
+* Cover `<FindImportProfileContainer>` container with tests (UIPFIMP-44)
+* Update 'source-manager-job-executions' dependency versions (UIPFIMP-45)
 
 ## [5.0.0](https://github.com/folio-org/ui-plugin-find-import-profile/tree/v5.0.0) (2021-10-08)